Solved problem 45. Refactored GetTriangularNumber method and added triangular, pentagonal and hexagonal number checks
git-tfs-id: [https://kaczmar2.visualstudio.com/DefaultCollection]$/Euler Project;C25

+++ EulerProject/EulerProject/Problems/Problem12.cs
@@ -15,7 +15,7 @@ namespace EulerProject.Problems
 
             do
             {
-                tn = Common.GetTriangularNumber(num);
+                tn = (long) Common.GetTriangularNumber(num);
                 numDivisors = Common.GetDivisorsCount(tn);
                 num++;
             } while (numDivisors <= target);
